Move DataObject::setData to MutableDataObjectTrait::setData
